requirements.txt
fastapi==0.70.0
mongoengine==0.23.1
pydantic==1.8.2
strawberry-graphql[fastapi]==0.85.1
uvicorn==0.15.0
